refactor: designPresets 관리자 확인 추가

+++ src/designPresets/designPresets.controller.ts
@@ -1,4 +1,15 @@
-import { Controller, Get, Post, Put, Delete, Param, Body, Request, UseGuards } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Put,
+  Delete,
+  Param,
+  Body,
+  Request,
+  UseGuards,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
 import { DesignPresetsService } from './designPresets.service';
 import { CreateDesignPresetDto } from './dto/create-designPresets.dto';
@@ -12,14 +23,22 @@ export class DesignPresetsController {
 
   @Get()
   async getDesignPresets(@Request() req): Promise<DesignPreset[]> {
-    const restaurantId = req.user.restaurantId;
+    if (req.user.type === 'login') {
+      const restaurantId = req.user.restaurantId;
 
-    return this.designPresetsService.getDesignPresets(restaurantId);
+      return this.designPresetsService.getDesignPresets(restaurantId);
+    }
+
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 
   @Get(':designPreset_id')
-  async getDesignPresetById(@Param('designPreset_id') designPresetId: string): Promise<DesignPreset> {
-    return this.designPresetsService.getDesignPresetById(designPresetId);
+  async getDesignPresetById(@Request() req, @Param('designPreset_id') designPresetId: string): Promise<DesignPreset> {
+    if (req.user.type === 'login') {
+      return this.designPresetsService.getDesignPresetById(designPresetId);
+    }
+
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 
   @Post()
@@ -27,21 +46,34 @@ export class DesignPresetsController {
     @Request() req,
     @Body() createDesignPresetDto: CreateDesignPresetDto,
   ): Promise<DesignPreset> {
-    const restaurantId = req.user.restaurantId;
+    if (req.user.type === 'login') {
+      const restaurantId = req.user.restaurantId;
 
-    return this.designPresetsService.createDesignPreset(restaurantId, createDesignPresetDto);
+      return this.designPresetsService.createDesignPreset(restaurantId, createDesignPresetDto);
+    }
+
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 
   @Put(':designPreset_id')
   async updateDesignPreset(
+    @Request() req,
     @Param('designPreset_id') designPresetId: string,
     @Body() updateDesignPresetDto: UpdateDesignPresetDto,
   ): Promise<DesignPreset> {
-    return this.designPresetsService.updateDesignPreset(designPresetId, updateDesignPresetDto);
+    if (req.user.type === 'login') {
+      return this.designPresetsService.updateDesignPreset(designPresetId, updateDesignPresetDto);
+    }
+
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 
   @Delete(':designPreset_id')
-  async deleteDesignPreset(@Param('designPreset_id') designPresetId: string): Promise<void> {
-    return this.designPresetsService.deleteDesignPreset(designPresetId);
+  async deleteDesignPreset(@Request() req, @Param('designPreset_id') designPresetId: string): Promise<void> {
+    if (req.user.type === 'login') {
+      return this.designPresetsService.deleteDesignPreset(designPresetId);
+    }
+
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 }
